+++ src/Dialect/ONNX/CMakeLists.txt
@@ -40,12 +40,14 @@ add_onnx_mlir_library(OMONNXOps
   ONNXOps/Math/ElementwiseBroadcast.cpp
   ONNXOps/Math/ElementwiseUnary.cpp
   ONNXOps/Math/Gemm.cpp
+  ONNXOps/Math/HardMax.cpp
   ONNXOps/Math/IsNan.cpp
   ONNXOps/Math/LRN.cpp
   ONNXOps/Math/MatMul.cpp
   ONNXOps/Math/RandomNormal.cpp
   ONNXOps/Math/RandomNormalLike.cpp
   ONNXOps/Math/Reduction.cpp
+  ONNXOps/Math/Scatter.cpp
   ONNXOps/Math/TopK.cpp
   ONNXOps/NN/Conv.cpp
   ONNXOps/NN/Dropout.cpp
