+++ src/Dialect/ONNX/ONNXOps/OpHelper.hpp
@@ -276,5 +276,5 @@ int64_t mlirTypeToOnnxType(mlir::Type elemType);
 #define NOT_IMPLEMENTED_MESSAGE                                                \
   (getOperationName() +                                                        \
       ": is not supported at this time. Please open an issue on "              \
-      "https://github.com/onnx/onnx-mlir and/or consider contribute code. "    \
+      "https://github.com/onnx/onnx-mlir and/or consider contributing code. "  \
       "Error encountered in shape inference.")
